+++ SharePointFramework/PortfolioWebParts/src/components/ResourceAllocation/index.tsx
@@ -16,6 +16,7 @@ import { IColumn } from 'office-ui-fabric-react/lib/DetailsList'
 import { MessageBar, MessageBarType } from 'office-ui-fabric-react/lib/MessageBar'
 import { format } from 'office-ui-fabric-react/lib/Utilities'
 import * as strings from 'PortfolioWebPartsStrings'
+import { UserMessage } from 'pp365-shared/lib/components/UserMessage'
 import { tryParsePercentage } from 'pp365-shared/lib/helpers'
 import { DataSourceService } from 'pp365-shared/lib/services'
 import React, { Component } from 'react'
@@ -85,11 +86,15 @@ export class ResourceAllocation extends Component<
 
   public render(): React.ReactElement<IResourceAllocationProps> {
     if (this.state.loading) return null
+
     if (this.state.error) {
       return (
         <div className={styles.root}>
           <div className={styles.container}>
-            <MessageBar messageBarType={MessageBarType.error}>{this.state.error}</MessageBar>
+            <UserMessage
+              text={this.state.error}
+              type={MessageBarType.success}
+            />
           </div>
         </div>
       )
@@ -282,9 +287,14 @@ export class ResourceAllocation extends Component<
     const groupNames: string[] = searchResults
       .map((res) => {
         const name = res.RefinableString71 || `${res.RefinableString72}|R`
+        if (name === null)
+          this.setState({
+            error: format(strings.ResourceAllocationErrorTransformGroupText, res.SiteTitle)
+          })
         return name
       })
       .filter((value, index, self) => self.indexOf(value) === index)
+
     let groups: ITimelineGroup[] = groupNames.map((name, id) => {
       const [title, type] = name.split('|')
       return {