+++ SharePointFramework/PortfolioWebParts/src/loc/mystrings.d.ts
@@ -77,6 +77,7 @@ declare interface IPortfolioWebPartsStrings {
   DeliveryStatusCommentLabel: string
   DeliveryStatusLabel: string
   DescriptionFieldLabel: string
+  TagFieldLabel: string
   DesiredValueLabel: string
   EditViewText: string
   ErrorText: string
@@ -120,6 +121,7 @@ declare interface IPortfolioWebPartsStrings {
   ProjectTimelineInfoText: string
   ProjectTimelineErrorTransformItemText: string
   ResourceAllocationInfoText: string
+  ResourceAllocationErrorTransformGroupText: string
   ResourceLabel: string
   ResponsibleLabel: string
   RoleLabel: string