+++ SharePointFramework/ProjectWebParts/src/components/ProjectTimeline/types.ts
@@ -89,6 +89,7 @@ export interface IItemData {
   hexColor?: string,
   elementType?: string,
   filter?: boolean
+  tag?: string
 }
 
 export interface ITimelineItem {